+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/reporter/CheckstyleReporterTest.groovy
@@ -22,9 +22,7 @@ class CheckstyleReporterTest extends AbstractKitTest {
 
             task testReport() {
                 doLast {
-                    new ru.vyarus.gradle.plugin.quality.report.CheckstyleReporter(
-                        new ru.vyarus.gradle.plugin.quality.ConfigLoader(project)
-                    ).report(checkstyleMain, 'main')
+                    new ru.vyarus.gradle.plugin.quality.report.CheckstyleReporter().report(checkstyleMain, 'main')
                 }
             }
         """)